Updated with filter options in addon Preferences

+++ cto_purge.py
@@ -20,7 +20,7 @@
 bl_info = {"name": "CTO Purge",
            "description": "Purge all custom transform orientations",
            "author": "Marcelo M. Marques (based on script by 'iyadahmed' Iyad Ahmed)",
-           "version": (1, 0, 1),
+           "version": (1, 0, 2),
            "blender": (2, 80, 75),
            "location": "View3D > Transform Orientations panel",
            "support": "COMMUNITY",
@@ -31,6 +31,10 @@ bl_info = {"name": "CTO Purge",
 
 # --- ### Change log
 
+# v1.0.2 (09.05.2022) - by Marcelo M. Marques
+# Added: configuration options to filter out custom transform orientations
+# Fixed: issue caused by special characters in the names
+
 # v1.0.1 (11.11.2021) - by Marcelo M. Marques
 # Chang: just minor clean up
 
@@ -39,6 +43,9 @@ bl_info = {"name": "CTO Purge",
 
 # --- ### Imports
 import bpy
+import os
+from bpy_extras.io_utils import ImportHelper
+
 
 class CTO_OT_Purge(bpy.types.Operator):
     """ This is a workaround for this issue:
@@ -46,9 +53,32 @@ class CTO_OT_Purge(bpy.types.Operator):
     """
     bl_idname = "object.cto_purge"
     bl_label = "Purge Custom Orientations  "
-    bl_description = "Purge all custom transform orientations"
+    bl_description = "Purge all custom transform orientations.\nCheck addon preferences for filter out options"
     bl_options = {"REGISTER"}
 
+    def CTO_FILTER(self):
+        """ Indicates which character(s) must be considered as a filter out option. """
+        try:
+            if __package__.find(".") != -1:
+                package = __package__[0:__package__.find(".")]
+            else:
+                package = __package__
+            strFilter = bpy.context.preferences.addons[package].preferences.CTO_FILTER
+        except Exception as e:
+            strFilter = ""
+        return (strFilter)
+
+    def CTO_SPACE(self):
+        """ Indicates if SPACE character must also be considered as a filter out option. """
+        try:
+            if __package__.find(".") != -1:
+                package = __package__[0:__package__.find(".")]
+            else:
+                package = __package__
+            bolSpace = bpy.context.preferences.addons[package].preferences.CTO_SPACE
+        except Exception as e:
+            bolSpace = False
+        return (bolSpace)
 
     def execute(self, context):
         # Try to set transform orientation and catch error message
@@ -56,15 +86,28 @@ class CTO_OT_Purge(bpy.types.Operator):
             bpy.context.scene.transform_orientation_slots[0].type = ""
         except Exception as inst:
             # Extract a list of transform orientations from error message
-            transforms = str(inst).split("in")[1][3:-2].replace("', '", " ").split()
+            # if we tried to set transform_orientation_slots[0].type to a non existent value, we get error message similar to the one below:
+            # > TypeError: bpy_struct: item.attr = val: enum "" not found in ('GLOBAL', 'LOCAL', 'NORMAL', 'GIMBAL', 'VIEW', 'CURSOR', 'Cube', 'Cube.001', 'Cube.002')
+            # where each element from the 7th position forward corresponds to one of the custom transform orientations (in the above example: 'Cube', 'Cube.001' and 'Cube.002')
+            transforms = str(inst).split("not found in")[1][3:-2].replace("', '", ",").split(",")
+            strFilters = self.CTO_FILTER().strip().upper()
+            if self.CTO_SPACE():
+                strFilters += " "
             # Exclude first 6 "default" transform orientations
             for type in transforms[6:]:
-                try:
-                    # Try to delete each custom transform orientation
-                    bpy.context.scene.transform_orientation_slots[0].type = type
-                    bpy.ops.transform.delete_orientation()
-                except Exception as e:
-                    pass
+                purge = True
+                for char in strFilters:
+                    # Will filter out (skip) any row that has any of the filter characters in its name
+                    if type.upper().find(char) >= 0:
+                        purge = False
+                        break
+                if purge:
+                    # Try to delete this custom transform orientation
+                    try:
+                        bpy.context.scene.transform_orientation_slots[0].type = type
+                        bpy.ops.transform.delete_orientation()
+                    except Exception as e:
+                        pass
         return {'FINISHED'}
 
 
@@ -77,7 +120,7 @@ class CTO_OT_Dummy(bpy.types.Operator):
     @classmethod
     def poll(cls, context):
         return False
-        
+
     def execute(self, context):
         return {'FINISHED'}
 
@@ -103,4 +146,3 @@ def unregister():
 
 if __name__ == '__main__':
     register()
-
